refactor: rename page directories to camel case
