+++ src/pages/FollowUpAnalysis/follow-up-analysis-table/FollowUpAnalysisTable.module.css
