+++ src/pages/OutlierDetection/constants.js
