+++ src/pages/ValidationRulesAnalysis/ValidationRulesDetails/ValidationRulesDetails.js
